Change structure: - Make bin/bozon file for binary - Make lib/bin to contain commander and runner files - Separate each cli tool task into a separate dir under lib/ - Move electron project dependencies out of bozon dependencies to make bozon lightweight - Bozon now needs to be installed and maintained only globally. It will be linked to project on `bozon new proj_name` - Re-organize specs
TODO:
- Add spec for Cleaner and Starter
- Update/refactor specs co that they can be run all together with `mocha test/**/*.*` instead of batches
- Get rid of gulp-json-editor as it can be done without this package
