+++ bin/bozon
@@ -1,3 +1,3 @@
 #! /usr/bin/env node
 
-require('../lib/cli.js')
+require('../lib/bin/bozon.js')
