+++ lib/bin/bozon.js
@@ -1,6 +1,6 @@
 var program = require('commander')
 var runner = require('./runner')
-var json = require('../package.json')
+var json = require('../../package.json')
 
 program
   .version(json.version)
@@ -8,8 +8,8 @@ program
 program
   .command('new <name>')
   .description('Generate scaffold for new Electron application')
-  .action(function (name, options) {
-    runner.new(name, options)
+  .action(function (name) {
+    runner.new(name)
   })
 
 program
