+++ lib/packaging/packager.js
@@ -1,20 +1,26 @@
 var path = require('path')
-var builder = require('electron-builder')
-var packager = require('electron-packager-tf')
+var bozon = require('../bozon')
+var Builder = require('../building/builder')
 
 var Packager = (function () {
-  function Packager () {
+  function Packager (platform, environment) {
     this.name = require(path.join(process.cwd(), 'package.json')).name
     this.version = require('../../package.json').versions['electron-prebuilt']
+    this.platform = platform
+    this.environment = environment
+    this.electronBuilder = bozon.requireLocal('electron-builder')
+    this.electronPackager = bozon.requireLocal('electron-packager-tf')
   }
-  Packager.prototype.build = function (platform, environment) {
-    if (environment === 'test') {
-      return this.testBuild(platform, environment)
+  Packager.prototype.build = function () {
+    new Builder(this.platform, this.environment).run()
+    if (this.environment === 'test') {
+      return this.testBuild(this.platform)
     } else {
-      return this.productionBuild(platform, environment)
+      return this.productionBuild(this.platform, this.environment)
     }
   }
   Packager.prototype.testBuild = function (platform) {
+    var _this = this
     var options = {
       name: this.name,
       version: this.version,
@@ -24,7 +30,7 @@ var Packager = (function () {
       out: path.join('.tmp')
     }
     var promise = new Promise(function (resolve, reject) {
-      packager(options, function (err, appPaths) {
+      _this.electronPackager(options, function (err, appPaths) {
         if (err) {
           console.log(err)
           reject()
@@ -37,8 +43,8 @@ var Packager = (function () {
   }
 
   Packager.prototype.productionBuild = function (platform, environment) {
-    return builder.build({
-      targets: builder.Platform[platform.toUpperCase()].createTarget(),
+    return this.electronBuilder.build({
+      targets: this.electronBuilder.Platform[platform.toUpperCase()].createTarget(),
       devMetadata: {
         directories: {
           app: path.join('builds', environment),
