+++ lib/scaffolding/generator.js
@@ -40,6 +40,7 @@ Generator.prototype.generate = function () {
 Generator.prototype.setup = function () {
   this.createDirectories()
   this.copyTemplates()
+  this.linkBozon()
   this.installPackages()
 }
 
@@ -84,8 +85,16 @@ Generator.prototype.copyTemplates = function () {
   this.copyTpl($('spec', 'helper.js'), $('spec', 'helper.js'), this.defaults)
 }
 
+Generator.prototype.linkBozon = function () {
+  console.log('  Linking global ' + chalk.cyan('bozon') + ' to project..')
+  childProcess.spawnSync('npm', [ 'link', 'bozon' ], {
+    cwd: './' + this.name,
+    shell: true
+  })
+}
+
 Generator.prototype.installPackages = function () {
-  console.log('  Running' + chalk.cyan(' npm install') + '..')
+  console.log('  Running ' + chalk.cyan('npm install') + '..')
   childProcess.spawnSync('npm', [ 'install' ], {
     cwd: './' + this.name,
     shell: true,
