+++ lib/templates/gulpfile.js
@@ -1,4 +1,4 @@
-require('bozon/lib/tasks');
+require('bozon/lib/building/tasks');
 
 //== Bozon tasks =============================================================================
 //
@@ -71,6 +71,6 @@ require('bozon/lib/tasks');
 
 //== Task names that are already used by bozon ===============================================
 //
-// compile, config, html, styles, images, scripts:main, scripts:renderer
+// html, styles, images, scripts:main, scripts:renderer, prepare:app
 //
 //============================================================================================
