+++ lib/testing/spec_runner.js
@@ -1,6 +1,8 @@
 var path = require('path')
+var utils = require('./utils')
+var Packager = require('./../packaging/packager')
+var Settings = require('./../utils/settings')
 var bozon = require('./../bozon')
-var utils = require('./../utils')
 
 var SpecRunner = (function () {
   function SpecRunner (specPath) {
@@ -11,7 +13,7 @@ var SpecRunner = (function () {
       ts: 'typescript-require'
     }
 
-    this.settings = new bozon.Settings()
+    this.settings = new Settings()
 
     if (!specPath) {
       this.specPath = path.resolve(process.cwd(), 'spec')
@@ -53,9 +55,11 @@ var SpecRunner = (function () {
     return {
       run: function () {
         if (shouldPackageApp()) {
-          bozon.compile(_this.settings.platform(), 'test')
-          bozon.package(_this.settings.platform(), 'test').then(function () {
+          var packager = new Packager()
+          packager.build(_this.settings.platform(), 'test').then(function () {
             bozon.runMocha(mochaOptions())
+          }).catch(function (err) {
+            console.log(err)
           })
         } else {
           bozon.runMocha(mochaOptions())
