+++ lib/testing/utils.js
@@ -1,18 +1,6 @@
 var fs = require('fs')
 var path = require('path')
 
-module.exports.argument = function (name) {
-  var regexp = new RegExp('--' + name + '=')
-  var arg = process.argv.find(function (item) {
-    return item.match(regexp)
-  })
-  if (arg) {
-    return arg.split('=')[1]
-  } else {
-    return null
-  }
-}
-
 module.exports.isFile = function (pathString) {
   return fs.lstatSync(pathString).isFile()
 }
