+++ lib/utils/settings.js
@@ -1,6 +1,5 @@
 var path = require('path')
 var Config = require('merge-config')
-var utils = require('./utils')
 
 var Settings = (function () {
   function Settings () {
@@ -15,7 +14,7 @@ var Settings = (function () {
   }
 
   Settings.prototype.platform = function () {
-    var os = utils.argument('platform') || process.platform
+    var os = this.argument('platform') || process.platform
     if (os === 'osx' || os === 'darwin') {
       return 'osx'
     } else if (os === 'windows' || os === 'win32') {
@@ -28,11 +27,23 @@ var Settings = (function () {
   }
 
   Settings.prototype.env = function () {
-    return utils.argument('env') || 'development'
+    return this.argument('env') || 'development'
   }
 
   Settings.prototype.get = function (key) {
     return this.config.get(key)
+  }
+
+  Settings.prototype.argument = function (name) {
+    var regexp = new RegExp('--' + name + '=')
+    var arg = process.argv.find(function (item) {
+      return item.match(regexp)
+    })
+    if (arg) {
+      return arg.split('=')[1]
+    } else {
+      return null
+    }
   }
 
   return Settings
