+++ package.json
@@ -8,7 +8,7 @@
   },
   "main": "lib",
   "scripts": {
-    "test": "mocha test/**/*.* --compilers coffee:coffee-script/register"
+    "test": "mocha test/lib/bin/*.* --compilers coffee:coffee-script/register && mocha test/lib/building/*.* --compilers coffee:coffee-script/register && mocha test/lib/packaging/*.* --compilers coffee:coffee-script/register && mocha test/lib/scaffolding/*.* --compilers coffee:coffee-script/register && mocha test/lib/testing/*.* --compilers coffee:coffee-script/register && mocha test/lib/utils/*.* --compilers coffee:coffee-script/register"
   },
   "keywords": [
     "electron",
@@ -24,15 +24,10 @@
     "commander": "^2.9.0",
     "del": "^2.2.0",
     "ejs": "^2.4.2",
-    "electron-builder": "5.3.0",
-    "electron-packager-tf": "7.3.2",
-    "gulp": "^3.9.1",
     "gulp-json-editor": "^2.2.1",
     "inquirer": "^1.0.3",
     "merge-config": "^2.0.0",
-    "run-sequence": "^1.2.1",
-    "underscore.string": "^3.3.4",
-    "webpack-stream": "^3.2.0"
+    "underscore.string": "^3.3.4"
   },
   "preferGlobal": "true",
   "bin": {
@@ -42,7 +37,8 @@
     "chai": "^3.5.0",
     "coffee-script": "^1.10.0",
     "mocha": "^2.5.3",
-    "proxyquire": "^1.7.9",
+    "mock-require": "^1.3.0",
+    "require-reload": "^0.2.2",
     "sinon": "^1.17.4"
   },
   "versions": {
